Search Bar Update
The search bar works now! Just start typing in it and take a look.

There is a bug, though, but we probably don’t need to worry about it
before we turn in the project. If you search for something, you can
delete lines from the resultant list of images that match your search.
It doesn’t actually delete the images from the playlist you’re in. If
you look at the library or the playlist you’re searching it, the image
you supposedly deleted is still there. In order to fix it, a LOT of
code will need to be rewritten, so I’m just not going to worry about it
for now.