Imgscalr Implementation
This update implements the external library “imgscalr” to handle
scaling the images displayed in the DisplayPane on the right side of
the screen. Try resizing the pane or resizing the program window! At
the moment, the DisplayPane shows “MonaLisa.jpg” by default. You will
need to change the path to MonaLisa.jpg in DisplayPane.java to whatever
the absolute path is to that file on your computer in order to get the
program to read it. Otherwise, instead of displaying the picture, the
DisplayPane will display an error message.