Committing and Saving Open Issue #2
This is the commit that fixes Issue #2, “Quitting and Saving”. This
commit was written by Michele Gravel, but I had to revert her changes
to the master branch and stick it in its own branch to merge changes
from there because other branches were getting stale. I know for a fact
that there is a better way to do this, but I don’t actually know how to
do it, so I’m doing it this way.