Adding Images, fixes #8, fixes #9
Wooooo! We can add our own image files to the library now! Click the
green “Plus” button to open the JFileChooser window and select one or
more image files to open. It’ll add them all to the library.

Then select an image in the table. It displays that image in the
DisplayPane!!!