Editing table cells, fixes #22
We can edit the cells in the JTable now, and doing so will update the
information stored in the ImageObject.

I also made a few tweaks. Resizing the window now gives priority to the
additional space to the JTable instead of the DisplayPane. I added
columns to the JTable for Date Added and page number. I also tweaked
the adjustColumnWidth method of the JTable so now it adjusts the
columns to reasonable widths based on the data in the column.