TableRowSorter and Slideshow, fixes #6, #21
I officially declare that SunPig is ready to be turned in for grading!
Changes include:

1. You can click the column headers to sort the table by that column in
ascending order. Click it again to sort descending.

2. Jessica wrote the actionListeners for the “<<”, “>”, and “>>”
buttons. I tweaked it so that when you click the play button, it turns
in to a pause button, and when you click it again, it turns back into a
play button.

3. Eventually I managed to get the slideshow to work when you click the
play button. It turned out to be waaaay more complicated than I
originally realized. I ended up having to learn multithreading, and run
the slideshow on a separate thread.

4. The sidebar is a little different now. The Image Library is now
separated from the playlists, and the playlists are indented.

5. I got rid of a whole bunch of extraneous code that I’d written when
I first made the ImageList classes, but that we don’t use anymore
because we do it better elsewhere.

6. I made a bunch of comments throughout the code to explain stuff.